refactor: [US-055] - Drop dna_analyser shim; point all importers at core.services.dna
